WIP, hack to disable full Android background location mode
We don't want this, we just want 'while in use' mode with the notification and foreground service to keep location updating while the user is in another app temporarily. Unfortunately no option to ask for this has been implemented; let's try this out for now and then talk to upstream.